pyproject.toml
[project]
name = "naloga"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "black>=25.1.0",
    "cairosvg>=2.7.1",
    "pillow>=11.2.0",
    "pydantic>=2.11.1",
    "requests>=2.32.3",
    "torch>=2.6.0",
    "transformers>=4.50.3",
]

[tool.black]
line-length = 120
